+++ doc/includes/references.h
@@ -1211,8 +1211,9 @@ struct mjModel_ {
   int*      tex_type;             // texture type (mjtTexture)                (ntex x 1)
   int*      tex_height;           // number of rows in texture image          (ntex x 1)
   int*      tex_width;            // number of columns in texture image       (ntex x 1)
-  int*      tex_adr;              // address in rgb                           (ntex x 1)
-  mjtByte*  tex_rgb;              // rgb (alpha = 1)                          (ntexdata x 1)
+  int*      tex_nchannel;         // number of channels in texture image      (ntex x 1)
+  int*      tex_adr;              // start address in tex_data                (ntex x 1)
+  mjtByte*  tex_data;             // pixel values                             (ntexdata x 1)
   int*      tex_pathadr;          // address of texture asset path; -1: none  (ntex x 1)
 
   // materials