+++ include/mujoco/mjxmacro.h
@@ -410,8 +410,9 @@
     X   ( int,     tex_type,              ntex,          1                    ) \
     X   ( int,     tex_height,            ntex,          1                    ) \
     X   ( int,     tex_width,             ntex,          1                    ) \
+    X   ( int,     tex_nchannel,          ntex,          1                    ) \
     X   ( int,     tex_adr,               ntex,          1                    ) \
-    X   ( mjtByte, tex_rgb,               ntexdata,      1                    ) \
+    X   ( mjtByte, tex_data,              ntexdata,      1                    ) \
     XMJV( int,     tex_pathadr,           ntex,          1                    ) \
     XMJV( int,     mat_texid,             nmat,          mjNTEXMAT            ) \
     XMJV( mjtByte, mat_texuniform,        nmat,          1                    ) \